moving function to the file where it's used
